# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aboubakry Dieng Ges-Compte API",
        "description": "Documentation de l'API Ges-Compte développée par Aboubakry Dieng",
        "contact": {
            "name": "Aboubakry Dieng",
            "email": "aboubakry.dieng@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gescompte-1.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lister toutes les transactions",
                "operationId": "0ffe6bf7980513c9d322e1ba4d400ff2",
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Créer une nouvelle transaction",
                "operationId": "b0ee1d82e907b370897f7c2adde8a664",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "montant",
                                    "compte_id"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "depot",
                                            "retrait",
                                            "virement",
                                            "paiement"
                                        ]
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "beneficiaire": {
                                        "type": "string"
                                    },
                                    "date_transaction": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "compte_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Obtenir une transaction spécifique",
                "operationId": "c9e072c8c645241005e971a691e8338e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Mettre à jour une transaction",
                "operationId": "7be6f226f094053033fa504c693d257c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "depot",
                                            "retrait",
                                            "virement",
                                            "paiement"
                                        ]
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "beneficiaire": {
                                        "type": "string"
                                    },
                                    "date_transaction": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Supprimer une transaction",
                "operationId": "db289b0d1fea21825cc577ddeb7f2157",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction supprimée"
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "title": "Transaction",
                "description": "Représente une transaction bancaire",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reference": {
                        "type": "string",
                        "example": "TRX20251026001"
                    },
                    "compte_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "depot"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    },
                    "description": {
                        "type": "string",
                        "example": "Dépôt salaire"
                    },
                    "beneficiaire": {
                        "type": "string",
                        "example": "Jean Martin"
                    },
                    "date_transaction": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}